+++ src/main/java/com/ivanbasic/learnspring/db2/repo/DepartmentsRepo.java
@@ -1,0 +1,7 @@
+package com.ivanbasic.learnspring.db2.repo;
+
+import com.ivanbasic.learnspring.db2.model.Departments;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface DepartmentsRepo extends JpaRepository<Departments, Integer> {
+}