+++ src/test/java/com/ivanbasic/learnspring/Db2HrControllerTest.java
@@ -1,0 +1,33 @@
+package com.ivanbasic.learnspring;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.MediaType;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+
+import static org.hamcrest.Matchers.containsString;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@SpringBootTest
+@AutoConfigureMockMvc
+@ActiveProfiles( "test" )
+public class Db2HrControllerTest {
+
+    @Autowired
+    private MockMvc mvc;
+
+    @Test
+    public void testTest() throws Exception {
+        mvc.perform(MockMvcRequestBuilders.get("/db2/hr").accept(MediaType.APPLICATION_JSON))
+                .andDo(print())
+                .andExpect(status().isOk())
+                .andExpect(content().string( containsString(  "name,Ada=1" )));
+    }
+
+}